+++ core/modules/modVereine.class.php
@@ -52,7 +52,7 @@ class modVereine extends DolibarrModules
 		$this->descriptionlong = 'ModuleVereineDescLong';
 		$this->editor_name = 'IT-Tabelander';
 		$this->editor_url = 'https://it.tabelander.co.at';
-		$this->version = '0.4.0-beta';
+		$this->version = '0.4.1-beta';
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		$this->picto = 'fa-landmark';
 
@@ -298,7 +298,7 @@ class modVereine extends DolibarrModules
 		// The functions suggested for Austria; changed or switched off ones stay as the association left them.
 		dol_include_once('/vereine/class/vereinefunctions.class.php');
 		$functions = new VereineFunctions($this->db);
-		if ($functions->ensureStandard() < 0) {
+		if ($functions->ensureStandard() < 0 || $functions->ensureFields() < 0) {
 			$this->error = $functions->error;
 			dol_syslog('modVereine::init '.$functions->error, LOG_ERR);
 		}